[ADJ] Complete the project
1. adjust Diet object
2. add db controller of diet page
3. adjust rule when access into the app

+++ app/src/main/java/com/arctic/aurora/Diet.java
@@ -2,9 +2,9 @@ package com.arctic.aurora;
 
 public class Diet {
     private final String name;
-    private final float calories;
+    private final int calories;
 
-    public Diet(String name, float calories) {
+    public Diet(String name, int calories) {
         this.name = name;
         this.calories = calories;
     }
@@ -12,8 +12,7 @@ public class Diet {
     public String getName() {
         return name;
     }
-
-    public float getCalories() {
+    public int getCalories() {
         return calories;
     }
 }
